+++ src/telliot/pricing/price_service.py
@@ -1,26 +1,12 @@
 from abc import ABC
 from abc import abstractmethod
-from datetime import datetime
 from typing import Any
 from typing import Dict
 from typing import Optional
 
 import requests
-from pydantic import BaseModel, Field
 
-
-class TimeStampedFloat(BaseModel):
-    """ A time-stamped floating point value
-
-    """
-    #: Value
-    val: float
-
-    #: Timestamp
-    ts: datetime = Field(default_factory=datetime.now)
-
-    def __init__(self, val, **data):
-        super().__init__(val=val, **data)
+from telliot.base import TimeStampedFloat
 
 
 class PriceServiceInterface(ABC):
