+++ tests/test_base.py
@@ -6,6 +6,7 @@ from pydantic.error_wrappers import ValidationError
 from telliot.base import Answer
 from telliot.base import TimeStampedAnswer
 from telliot.base import TimeStampedFixed
+from telliot.base import TimeStampedFloat
 
 
 def test_Answer():
@@ -32,4 +33,19 @@ def test_TimeStampedFixed_resolution():
     f = TimeStampedFixed(2.1234567)
     assert f.int == 2123457
     assert f.val == 2.123457
+
+
+def test_TimeStampedFloat():
+    obj = TimeStampedFloat(3.14)
+    assert obj.val == 3.14
+
+    obj = TimeStampedFloat('3.14')
+    assert obj.val == 3.14
+
+    obj = TimeStampedFloat(3.14, ts=datetime.now())
+
+    obj.ts = datetime.now()
+
+    with pytest.raises(TypeError):
+        _ = TimeStampedFloat()
